Added a command that gives you a random card respecting their percentage chance

+++ src/main/java/me/trololo11/lifespluginseason3/cardstuff/CardClass.java
@@ -57,7 +57,7 @@ public abstract class CardClass {
 
     public abstract CardType getCardType();
 
-    public abstract float getPercentageChance();
+    public abstract int getPercentageChance();
 
     public ItemStack getCardItem(){
